DEV: Introduce syntax_tree for ruby formatting

+++ lib/follow/updater.rb
@@ -11,9 +11,7 @@ class Follow::Updater
   end
 
   def unfollow
-    UserFollower
-      .where(follower_id: @follower.id, user_id: @target.id)
-      .destroy_all
+    UserFollower.where(follower_id: @follower.id, user_id: @target.id).destroy_all
   end
 
   private
@@ -21,58 +19,57 @@ class Follow::Updater
   def follow(notification_level)
     if @target.id == @follower.id
       raise Discourse::InvalidAccess.new(
-        nil,
-        nil,
-        custom_message: 'follow.user_cannot_follow_themself'
-      )
+              nil,
+              nil,
+              custom_message: "follow.user_cannot_follow_themself",
+            )
     end
 
-    %i(bot staged suspended).each do |status|
+    %i[bot staged suspended].each do |status|
       if @target.public_send(:"#{status}?")
         raise Discourse::InvalidAccess.new(
-          nil,
-          nil,
-          custom_message: "follow.user_cannot_follow_#{status}"
-        )
+                nil,
+                nil,
+                custom_message: "follow.user_cannot_follow_#{status}",
+              )
       end
     end
 
     if !Follow::Notification.levels.invert.key?(notification_level)
       raise Discourse::InvalidParameters.new(
-        I18n.t("follow.invalid_notification_level", level: notification_level.inspect)
-      )
+              I18n.t("follow.invalid_notification_level", level: notification_level.inspect),
+            )
     end
 
     if !@target.allow_people_to_follow_me
       raise Discourse::InvalidAccess.new(
-        nil,
-        nil,
-        custom_message: "follow.user_does_not_allow_follow",
-        custom_message_params: { username: @target.username }
-      )
+              nil,
+              nil,
+              custom_message: "follow.user_does_not_allow_follow",
+              custom_message_params: {
+                username: @target.username,
+              },
+            )
     end
 
     if @target.user_option&.hide_profile_and_presence
       raise Discourse::InvalidAccess.new(
-        nil,
-        nil,
-        custom_message: "follow.user_does_not_allow_follow",
-        custom_message_params: { username: @target.username }
-      )
+              nil,
+              nil,
+              custom_message: "follow.user_does_not_allow_follow",
+              custom_message_params: {
+                username: @target.username,
+              },
+            )
     end
 
-    relation = UserFollower.find_or_initialize_by(
-      user_id: @target.id,
-      follower_id: @follower.id,
-    )
+    relation = UserFollower.find_or_initialize_by(user_id: @target.id, follower_id: @follower.id)
     relation.level = notification_level
     relation.save!
 
     payload = {
       notification_type: Notification.types[:following],
-      data: {
-        display_username: @follower.username
-      }.to_json
+      data: { display_username: @follower.username }.to_json,
     }
     send_notification(payload) if should_notify?(payload)
 
@@ -80,10 +77,8 @@ class Follow::Updater
   end
 
   def should_notify?(payload)
-    SiteSetting.follow_notifications_enabled &&
-    @follower.notify_followed_user_when_followed &&
-    @target.notify_me_when_followed &&
-    !notification_sent_recently(payload)
+    SiteSetting.follow_notifications_enabled && @follower.notify_followed_user_when_followed &&
+      @target.notify_me_when_followed && !notification_sent_recently(payload)
   end
 
   def send_notification(payload)
@@ -91,6 +86,6 @@ class Follow::Updater
   end
 
   def notification_sent_recently(payload)
-    @target.notifications.where(payload).where('created_at >= ?', 1.day.ago).exists?
+    @target.notifications.where(payload).where("created_at >= ?", 1.day.ago).exists?
   end
 end
